+++ dchars/documentation/Sphinx/RSTsourcefiles/changelog.rst
@@ -3,6 +3,14 @@ CHANGELOG_TITLE
 ===============
 
 ------------------
+0.4.1 [2013_06_XX]
+------------------
+
+  * (bod) new function : ListOfInternalStructures.contains_unknown_character
+  * (bod) no more words stored in the buffers with unknown characters.
+  * (bod) words stored in the buffers are now stored with their real_indexes and indexes.
+
+------------------
 0.4.0 [2013_06_25]
 ------------------